Move no op snapshot adapter to internal package
